GUI: Implement a simple theming system and port Buttons to it :^)
The main object is a Theme, which is global to the whole app. Each
button type has its default colors which it takes from the theme. You
can override the colors by calling override_button_colors() and
modifying attributes you want to change.

The widgets that are currently themable are TextButtons, ImageButtons
and TabButtons. Datebox day buttons and MessageBox buttons have colors
hardcoded for now. This is left as a FIXME for now.