GUI: Avoid doing widget-related actions in constructors
Many Widget methods (like set_focused()) rely on WidgetTreeRoot be set
on widget, which cannot be guaranteed when object is just constructed.
This commit adds a new virtual function on_add(), which is called just
after WidgetTreeRoot is set to a widget, so that you can safely add
another widgets, which will inherit it from the parent.