EML: Fix memory leak when failing to dynamic cast object
We released the object from unique_ptr, then dynamic_casting which might
be null even if the argument was nonnull, therefore leaking the pointer.
This commit adds a `delete` so that the leak doesn't occur anymore, and
comments everything so that it can be seen from afar that this code
contains manual memory management.