EML: Don't call load_from_eml_object for every input object
Some loaders (especially these using require_property()) didn't work
because the property was checked both for class definition and
instantation, requiring user to define it in both places. Now the
objects are merged in the way that instantation overrides definition's
properties and load_from_eml_object() is called once for the merge
result.

This fixes using define with Container :^)