LLGL+EssaGUI/gfx: Start implementing Builder API
llgl::Builder and its descendants (the only is GUIBuilder for now) are
convenience classes to make high-level rendering much simpler. It is
meant to simplify tasks like "draw a rectangle", while still maintaining
performance by feeding the shapes to a single VAO.

The Gfx::GUIBuilder has functions for adding basic 2D shapes (currently
rectangle and regular polygon). It is meant to replace GUI::Window
draw_* functions in the future.