GUI/TextEditor: Add set_content_impl()
This function just sets content without updating anything. It can be
called by TextBox descendants to internally set the string while
avoiding recursion. Thanks to that, we can refactor normal set_content
so that it doesn't call on_change when NotifyUser is No as it was
before. This patch fixes strange ColorPicker behavior when sliding with
HSV slider sometimes.