LLGL+GUI: Introduce event target types
Previously checking which targets should receive an event were a mess
of is_global(), is_focused() and is_mouse_related() functions. Now it
is ordered into an EventTargetType enum which contains various target
types, which influence the procedure that is used in
is_affected_by_event().

This should remove need of manually doing is_hover() in Button and in
several places (only Button was removed for now).