GUI: Always use old focused overlay to send events to
Focused overlay may change in its handler, especially when this is a
button click that opens another overlay. Because of that, we have run
an event handler twice if old focused overlay is hovered: once because
it is focused, and once because it is hovered (and no longer focused, so
passes the check in line 116).

This commits makes the handler always use a well known (old) focused
overlay, so that the check in line 116 actually makes sense.

Also add a new debug option: GUI_DumpOverlayHandleEventCalls.