Refactor MDI subsystem
EssaGUI is no longer a window manager. Now it is like other GUI engines,
that is, we have a "root widget" for every Host Window. If you want
windows-in-windows, there is a GUI::MDI::Host widget for that.

For backwards compatibility (read: easier transition in Essa apps), we
now create that MDI Host as default widget. We can't drop this until
we convert all dialogs & special overlays to native windows.